services/svc-commerce/app/app/requirements.txt
fastapi
uvicorn[standard]
asyncpg
pydantic
python-jose[cryptography]
httpx
PyJWT>=2.8.0
argon2-cffi==23.1.0